Forgot to make moveto! a method of DAQCore.moveto!
Other functions such as move, which for now, are not methods of DAQCore (they don't actually exist!) has conflicts with other devices such as WMesa. One possibility is to add generic functions to DAQCore. Perhaps creating another package such as DAQMover or DAQRobot could be better. For now I will leave it as it is. If the package is used with anotrher one that has a function with the same name, just access it directly via RoboSimples.move for example.